LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-handler-throws-expected.txt

CONSOLE MESSAGE: TypeError: a message

Harness Error (FAIL), message = TypeError: a message

FAIL event.intercept() should abort if the handler throws assert_equals: expected "#1" but got ""
